+++ src/main/java/com/stoqing/reservas/entities/model/Estado.java
@@ -8,6 +8,7 @@ import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import lombok.ToString;
 
 @Entity
 @Table(name = "estado")
@@ -18,7 +19,7 @@ public class Estado {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_estado")
-    private Long id;
+    private int id;
 
     @NotBlank
     @Column(name = "nombre", nullable = false, length = 60)
